Refactor routing for testing and application
