+++ src/App.js
@@ -1,9 +1,9 @@
 import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import NavBar from './components/NavBar';
-import Home from './pages/Home';
+import Home from './components/Home';
 import GlobalStyles from './styles/GlobalStyles';
-import CryptoDetails from './pages/Detail';
+import CryptoDetails from './components/Detail';
 
 function App() {
   return (
