+++ src/components/Detail.js
